Move jinja templating tests alongside salt/utils/jinja.py's tests
