+++ tests/integration/files/file/base/templates/hello_import_generalerror
