+++ tests/integration/files/file/base/templates/hello_include
