+++ tests/integration/files/file/base/templates/macro
