+++ tests/integration/files/file/base/templates/macroerror
