+++ tests/integration/files/file/base/templates/macrogeneral
