+++ tests/integration/files/file/base/templates/non_ascii
